+++ vrpn_ForceDevice.C
@@ -238,7 +238,7 @@ vrpn_int32 vrpn_ForceDevice::decode_force (const char *buffer, const vrpn_int32 
     if (len !=  (3*sizeof(vrpn_float64)) ) {
       fprintf(stderr,"vrpn_ForceDevice: force message payload error\n");
       fprintf(stderr,"             (got %d, expected %lud)\n",
-		    len, 3*sizeof(vrpn_float64) );
+		    len, static_cast<unsigned long>(3*sizeof(vrpn_float64)) );
       return -1;
     }
 
@@ -283,7 +283,7 @@ vrpn_int32 vrpn_ForceDevice::decode_custom_effect (const char *buffer, const vrp
     if (static_cast<size_t>(len) < (sizeof(vrpn_uint32)*2)) {
 	  fprintf(stderr,"vrpn_ForceDevice: custom effect message payload error\n");
       fprintf(stderr,"             (got %d, expected at least %lud)\n",
-		    len, 2*sizeof(vrpn_uint32) );
+		    len, static_cast<unsigned long>(2*sizeof(vrpn_uint32)) );
       return -1;
     }
 
@@ -293,7 +293,7 @@ vrpn_int32 vrpn_ForceDevice::decode_custom_effect (const char *buffer, const vrp
     if ((vrpn_uint32)(len) < (2*sizeof(vrpn_uint32) + (*nbParams)*sizeof(vrpn_float32))) {
 	  fprintf(stderr,"vrpn_ForceDevice: custom effect message payload error\n");
       fprintf(stderr,"             (got %d, expected at least %lud)\n",
-		    len,(2*sizeof(vrpn_uint32) + (*nbParams)*sizeof(vrpn_float32)));
+		    len,static_cast<unsigned long>(2*sizeof(vrpn_uint32) + (*nbParams)*sizeof(vrpn_float32)));
       return -2;
     }
 
@@ -401,7 +401,7 @@ vrpn_int32 vrpn_ForceDevice::decode_plane(const char *buffer, const vrpn_int32 l
     if (len != 8*sizeof(vrpn_float32)+2*sizeof(vrpn_int32)){
 	    fprintf(stderr,"vrpn_ForceDevice: plane message payload error\n");
 	    fprintf(stderr,"             (got %d, expected %lud)\n",
-		    len, 8*sizeof(vrpn_float32)+2*sizeof(vrpn_int32) );
+		    len, static_cast<unsigned long>(8*sizeof(vrpn_float32)+2*sizeof(vrpn_int32)) );
 	    return -1;
     }
 
@@ -456,7 +456,7 @@ vrpn_int32 vrpn_ForceDevice::decode_surface_effects(const char *buffer, const vr
     if (len != 6*sizeof(vrpn_float32)){
         fprintf(stderr,"vrpn_ForceDevice: surface effects message payload ");
         fprintf(stderr,"error\n             (got %d, expected %lud)\n",
-		    len, 6*sizeof(vrpn_float32) );
+		    len, static_cast<unsigned long>(6*sizeof(vrpn_float32)) );
 	return -1;
     }
 
@@ -502,7 +502,7 @@ vrpn_int32 vrpn_ForceDevice::decode_vertex(const char *buffer,
     if (len != (sizeof(objNum) + sizeof(vertNum) + 3*sizeof(vrpn_float32))){
 	    fprintf(stderr,"vrpn_ForceDevice: vertex message payload error\n");
 	    fprintf(stderr,"             (got %d, expected %lud)\n",
-		    len, sizeof(objNum) +sizeof(vertNum) + 3*sizeof(vrpn_float32) );
+		    len, static_cast<unsigned long>(sizeof(objNum) +sizeof(vertNum) + 3*sizeof(vrpn_float32)) );
 	    return -1;
     }
 
@@ -546,7 +546,7 @@ vrpn_int32 vrpn_ForceDevice::decode_normal(const char *buffer,const vrpn_int32 l
     if (len != (sizeof(vrpn_int32) + sizeof(vrpn_int32) + 3*sizeof(vrpn_float32))){
 	    fprintf(stderr,"vrpn_ForceDevice: normal message payload error\n");
 	    fprintf(stderr,"             (got %d, expected %lud)\n",
-		    len, sizeof(vrpn_int32) + sizeof(vrpn_int32) + 3*sizeof(vrpn_float32) );
+		    len, static_cast<unsigned long>(sizeof(vrpn_int32) + sizeof(vrpn_int32) + 3*sizeof(vrpn_float32)) );
 	    return -1;
     }
 
@@ -596,7 +596,7 @@ vrpn_int32 vrpn_ForceDevice::decode_triangle(const char *buffer,
     if (len != (sizeof(vrpn_int32) + 7*sizeof(vrpn_int32))){
 	    fprintf(stderr,"vrpn_ForceDevice: triangle message payload error\n");
 	    fprintf(stderr,"             (got %d, expected %lud)\n",
-		    len, sizeof(vrpn_int32) + 7*sizeof(vrpn_int32) );
+		    len, static_cast<unsigned long>(sizeof(vrpn_int32) + 7*sizeof(vrpn_int32)) );
 	    return -1;
     }
 
@@ -639,7 +639,7 @@ vrpn_int32 vrpn_ForceDevice::decode_removeTriangle(const char *buffer,
     if (len != (sizeof(vrpn_int32) + sizeof(vrpn_int32))){
 	fprintf(stderr,"vrpn_ForceDevice: remove triangle message payload");
 	    fprintf(stderr," error\n             (got %d, expected %lud)\n",
-		    len, sizeof(vrpn_int32) +sizeof(vrpn_int32) );
+		    len, static_cast<unsigned long>(sizeof(vrpn_int32) +sizeof(vrpn_int32)) );
 	    return -1;
     }
 
@@ -684,7 +684,7 @@ vrpn_int32 vrpn_ForceDevice::decode_updateTrimeshChanges(const char *buffer,
     if (len != (sizeof(vrpn_int32) +4*sizeof(vrpn_float32))){
 	fprintf(stderr,"vrpn_ForceDevice: update trimesh message payload");
 	    fprintf(stderr," error\n             (got %d, expected %lud)\n",
-		    len, sizeof(vrpn_int32) +4*sizeof(vrpn_float32) );
+		    len, static_cast<unsigned long>(sizeof(vrpn_int32) +4*sizeof(vrpn_float32)) );
 	    return -1;
     }
 
@@ -725,7 +725,7 @@ vrpn_int32 vrpn_ForceDevice::decode_setTrimeshType(const char *buffer,const vrpn
     if (len != (sizeof(vrpn_int32) +sizeof(vrpn_int32))){
 	fprintf(stderr,"vrpn_ForceDevice: trimesh type message payload");
 	    fprintf(stderr," error\n             (got %d, expected %lud)\n",
-		    len, sizeof(vrpn_int32) +sizeof(vrpn_int32) );
+		    len, static_cast<unsigned long>(sizeof(vrpn_int32) + sizeof(vrpn_int32)) );
 	    return -1;
     }
 
@@ -765,7 +765,7 @@ vrpn_int32 vrpn_ForceDevice::decode_trimeshTransform(const char *buffer,
     if (len != (sizeof(vrpn_int32) +16*sizeof(vrpn_float32))){
 	fprintf(stderr,"vrpn_ForceDevice: trimesh transform message payload ");
 	    fprintf(stderr,"error\n             (got %d, expected %lud)\n",
-		    len, sizeof(vrpn_int32) +16*sizeof(vrpn_float32) );
+		    len, static_cast<unsigned long>(sizeof(vrpn_int32) + 16*sizeof(vrpn_float32)) );
 	    return -1;
     }
 
@@ -801,7 +801,7 @@ vrpn_int32 vrpn_ForceDevice::decode_addObject(const char *buffer,vrpn_int32 len,
     if (len != 2*sizeof(vrpn_int32) ){
 	fprintf(stderr,"vrpn_ForceDevice: add object message payload ");
 	    fprintf(stderr,"error\n             (got %d, expected %lud)\n",
-		    len, 2*sizeof(vrpn_int32) );
+		    len, static_cast<unsigned long>(2*sizeof(vrpn_int32)) );
 	    return -1;
     }
 
@@ -835,7 +835,7 @@ vrpn_int32 vrpn_ForceDevice::decode_addObjectExScene(const char *buffer,vrpn_int
     if (len != sizeof(vrpn_int32)){
 	fprintf(stderr,"vrpn_ForceDevice: add object message payload ");
 	    fprintf(stderr,"error\n             (got %d, expected %lud)\n",
-		    len, sizeof(vrpn_int32) );
+		    len, static_cast<unsigned long>(sizeof(vrpn_int32)) );
 	    return -1;
     }
 
@@ -872,7 +872,7 @@ vrpn_int32 vrpn_ForceDevice::decode_objectPosition(const char *buffer,vrpn_int32
     if (len != (sizeof(vrpn_int32) + 3*sizeof(vrpn_float32))){
 	fprintf(stderr,"vrpn_ForceDevice: object position message payload ");
 	    fprintf(stderr,"error\n             (got %d, expected %lud)\n",
-		    len, sizeof(vrpn_int32) + 3*sizeof(vrpn_float32) );
+		    len, static_cast<unsigned long>(sizeof(vrpn_int32) + 3*sizeof(vrpn_float32)) );
 	    return -1;
     }
 
@@ -912,7 +912,7 @@ vrpn_int32 vrpn_ForceDevice::decode_objectOrientation(const char *buffer,vrpn_in
     if (len != (sizeof(vrpn_int32) + 4*sizeof(vrpn_float32))){
 	fprintf(stderr,"vrpn_ForceDevice: object orientation message payload ");
 	    fprintf(stderr,"error\n             (got %d, expected %lud)\n",
-		    len, sizeof(vrpn_int32) + 4*sizeof(vrpn_float32) );
+		    len, static_cast<unsigned long>(sizeof(vrpn_int32) + 4*sizeof(vrpn_float32)) );
 	    return -1;
     }
 
@@ -952,7 +952,7 @@ vrpn_int32 vrpn_ForceDevice::decode_objectScale(const char *buffer,vrpn_int32 le
     if (len != (sizeof(vrpn_int32) + 3*sizeof(vrpn_float32))){
 	fprintf(stderr,"vrpn_ForceDevice: object scale message payload ");
 	    fprintf(stderr,"error\n             (got %d, expected %lud)\n",
-		    len, sizeof(vrpn_int32) + 3*sizeof(vrpn_float32) );
+		    len, static_cast<unsigned long>(sizeof(vrpn_int32) + 3*sizeof(vrpn_float32)) );
 	    return -1;
     }
 
@@ -988,7 +988,7 @@ vrpn_int32 vrpn_ForceDevice::decode_removeObject(const char *buffer,vrpn_int32 l
     if (len != sizeof(vrpn_int32) ){
 	fprintf(stderr,"vrpn_ForceDevice: remove object message payload ");
 	    fprintf(stderr,"error\n             (got %d, expected %lud)\n",
-		    len, sizeof(vrpn_int32) );
+		    len, static_cast<unsigned long>(sizeof(vrpn_int32)) );
 	    return -1;
     }
 
@@ -1021,7 +1021,7 @@ vrpn_int32 vrpn_ForceDevice::decode_clearTrimesh(const char *buffer,vrpn_int32 l
     if (len != sizeof(vrpn_int32) ){
 	fprintf(stderr,"vrpn_ForceDevice: clear TriMesh message payload ");
 	    fprintf(stderr,"error\n             (got %d, expected %lud)\n",
-		    len, sizeof(vrpn_int32) );
+		    len, static_cast<unsigned long>(sizeof(vrpn_int32)) );
 	    return -1;
     }
 
@@ -1055,7 +1055,7 @@ vrpn_int32 vrpn_ForceDevice::decode_moveToParent(const char *buffer,vrpn_int32 l
     if (len != (sizeof(vrpn_int32) +sizeof(vrpn_int32)) ){
 	fprintf(stderr,"vrpn_ForceDevice: move object to parent message payload ");
 	    fprintf(stderr,"error\n             (got %d, expected %lud)\n",
-		    len, sizeof(vrpn_int32) +sizeof(vrpn_int32) );
+		    len, static_cast<unsigned long>(sizeof(vrpn_int32) +sizeof(vrpn_int32)) );
 	    return -1;
     }
 
@@ -1097,7 +1097,7 @@ vrpn_int32 vrpn_ForceDevice::decode_setHapticOrigin(const char *buffer,vrpn_int3
     if (len != 7*sizeof(vrpn_int32)){
 	    fprintf(stderr,"vrpn_ForceDevice: sethapticorigin message payload error\n");
 	    fprintf(stderr,"             (got %d, expected %lud)\n",
-		    len, 7*sizeof(vrpn_int32) );
+		    len, static_cast<unsigned long>(7*sizeof(vrpn_int32)) );
 	    return -1;
     }
 	
@@ -1135,7 +1135,7 @@ vrpn_int32 vrpn_ForceDevice::decode_setHapticScale(const char *buffer,vrpn_int32
     if (len != sizeof(vrpn_float32)){
 	    fprintf(stderr,"vrpn_ForceDevice: sethapticscale message payload error\n");
 	    fprintf(stderr,"             (got %d, expected %lud)\n",
-		    len, 7*sizeof(vrpn_float32) );
+		    len, static_cast<unsigned long>(7*sizeof(vrpn_float32)) );
 	    return -1;
     }
 	
@@ -1176,7 +1176,7 @@ vrpn_int32 vrpn_ForceDevice::decode_setSceneOrigin(const char *buffer,vrpn_int32
     if (len != 7*sizeof(vrpn_int32)){
 	    fprintf(stderr,"vrpn_ForceDevice: setsceneorigin message payload error\n");
 	    fprintf(stderr,"             (got %d, expected %lud)\n",
-		    len, 7*sizeof(vrpn_int32) );
+		    len, static_cast<unsigned long>(7*sizeof(vrpn_int32)) );
 	    return -1;
     }
 	
@@ -1216,7 +1216,7 @@ vrpn_int32 vrpn_ForceDevice::decode_setObjectIsTouchable(const char *buffer,vrpn
     if (len != (sizeof(vrpn_int32)+sizeof(vrpn_bool))){
 	    fprintf(stderr,"vrpn_ForceDevice: set object is touchable message payload error\n");
 	    fprintf(stderr,"             (got %d, expected %lud)\n",
-		    len, sizeof(vrpn_int32)+sizeof(vrpn_bool) );
+		    len, static_cast<unsigned long>(sizeof(vrpn_int32)+sizeof(vrpn_bool)) );
 	    return -1;
     }
 
@@ -1266,7 +1266,7 @@ vrpn_int32 vrpn_ForceDevice::decode_forcefield(const char *buffer,
     if (len != 16*sizeof(vrpn_float32)){
        fprintf(stderr,"vrpn_ForceDevice: force field message payload error\n");
 	    fprintf(stderr,"             (got %d, expected %lud)\n",
-		    len, 16*sizeof(vrpn_float32) );
+		    len, static_cast<unsigned long>(16*sizeof(vrpn_float32)) );
 	    return -1;
     }
 
@@ -1312,7 +1312,7 @@ vrpn_int32 vrpn_ForceDevice::decode_error(const char *buffer,
     if (len != sizeof(vrpn_int32)){
 	    fprintf(stderr,"vrpn_ForceDevice: error message payload error\n");
 	    fprintf(stderr,"             (got %d, expected %lud)\n",
-		    len, sizeof(vrpn_int32) );
+		    len, static_cast<unsigned long>(sizeof(vrpn_int32)) );
 	    return -1;
     }
 
@@ -1396,7 +1396,7 @@ vrpn_int32 vrpn_ForceDevice::decode_enableConstraint
     fprintf(stderr,"vrpn_ForceDevice:  "
                    "enable constraint message payload error\n"
                    "             (got %d, expected %lud)\n",
-            len, sizeof(vrpn_int32));
+            len, static_cast<unsigned long>(sizeof(vrpn_int32)));
     return -1;
   }
   
@@ -1449,7 +1449,7 @@ vrpn_int32 vrpn_ForceDevice::decode_setConstraintMode
     fprintf(stderr,"vrpn_ForceDevice:  "
                    "constraint mode payload error\n"
                    "             (got %d, expected %lud)\n",
-            len, sizeof(vrpn_int32));
+            len, static_cast<unsigned long>(sizeof(vrpn_int32)));
     return -1;
   }
   
@@ -1578,7 +1578,7 @@ vrpn_int32 vrpn_ForceDevice::decode_setConstraintKSpring
     fprintf(stderr,"vrpn_ForceDevice:  "
                    "set constraint spring message payload error\n"
                    "             (got %d, expected %lud)\n",
-            len, sizeof(vrpn_float32));
+            len, static_cast<unsigned long>(sizeof(vrpn_float32)));
     return -1;
   }
 
@@ -1622,7 +1622,7 @@ vrpn_int32 vrpn_ForceDevice::decodePoint (const char * buffer,
     fprintf(stderr,"vrpn_ForceDevice:  "
                    "decode point message payload error\n"
                    "             (got size %d, expected %lud)\n",
-            len, 3 * sizeof(vrpn_float32));
+            len, static_cast<unsigned long>(3 * sizeof(vrpn_float32)));
     return -1;
   }
   
@@ -2483,7 +2483,7 @@ int vrpn_ForceDevice_Remote::handle_error_change_message(void *userdata,
     if (p.payload_len != sizeof(vrpn_int32)) {
 	    fprintf(stderr, "vrpn_ForceDevice: error message payload"
 		    " error\n(got %d, expected %lud)\n",
-		    p.payload_len, sizeof(vrpn_int32));
+		    p.payload_len, static_cast<unsigned long>(sizeof(vrpn_int32)));
 	    return -1;
     }
     tp.msg_time = p.msg_time;
